pyproject.toml
[project]
name = "promptseal-demo"
version = "0.1.0"
description = "PromptSeal hackathon demo: signed, hash-chained, anchored AI agent receipts."
requires-python = ">=3.11,<3.13"
dependencies = [
  "langchain==0.3.7",
  "langchain-anthropic==0.3.0",
  "langchain-core==0.3.21",
  "langchain-openai==0.2.10",
  "anthropic==0.39.0",
  "cryptography==43.0.3",
  "web3==7.5.0",
  "eth-account==0.13.4",
  "pydantic==2.9.2",
  "python-dotenv==1.0.1",
  "rich==13.9.0",
]

[dependency-groups]
dev = [
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0",
  "ruff==0.7.4",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["promptseal*", "agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
